+++ lib/feature/admission/kyiv_placement/placement_from_return/ui/widgets/nom_scan_dialog.dart
@@ -330,7 +330,7 @@ class _InputCountAlertState extends State<InputCountAlert> {
                 onPressed: () {
                   context.read<PlacementFromReturnCubit>().manualCountIncrement(
                       controller.text,
-                      widget.nom.qty.toDouble(),
+                      widget.nom.qty.toInt(),
                       widget.nom.count);
                   Navigator.pop(context);
                 },
@@ -518,7 +518,7 @@ class ChangeQuantity extends StatefulWidget {
       {super.key, required this.qty, required this.nom, required this.count});
 
   final int qty;
-  final double count;
+  final int count;
   final AdmissionNom nom;
 
   @override
@@ -578,7 +578,7 @@ class _ChangeQuantityState extends State<ChangeQuantity> {
           actions: [
             ElevatedButton(
                 onPressed: () {
-                  double inputCount = double.parse(controller.text);
+                  int inputCount = int.parse(controller.text);
                   if (controller.text.isNotEmpty) {
                     if (inputCount > widget.qty) {
                       Alerts(
