+++ lib/feature/inventory/inventory_by_cells/task_noms/cubit/task_noms_cubit.dart
@@ -80,7 +80,7 @@ class InventoryByCellsTaskNomsCubit
 
   void manualCountIncrement(String count) {
     emit(state.copyWith(
-      count: double.parse(count),
+      count: int.parse(count),
       status: InventoryStatus.success,
     ));
   }
