+++ lib/feature/inventory/inventory_by_nom/models/cells_by_nom.dart
@@ -28,7 +28,7 @@ class CellsByNom {
 class Cell {
   final String code;
   final String name;
-  final double planCount;
+  final int planCount;
   final int factCount;
   final String nomStatus;
